server/database/requirements.txt
psycopg2-binary
python-dotenv
supabase
